make IsFdConsumingAddr a standalone utility func
