+++ swarm_test.go
@@ -15,7 +15,6 @@ import (
 	"github.com/libp2p/go-libp2p-core/network"
 	"github.com/libp2p/go-libp2p-core/peer"
 	"github.com/libp2p/go-libp2p-core/peerstore"
-	"github.com/libp2p/go-libp2p-core/test"
 
 	. "github.com/libp2p/go-libp2p-swarm"
 	. "github.com/libp2p/go-libp2p-swarm/testing"
@@ -384,53 +383,6 @@ func TestConnectionGating(t *testing.T) {
 			}, 2*time.Second, 100*time.Millisecond, n)
 		})
 
-	}
-}
-
-func TestIsFdConsuming(t *testing.T) {
-	tcs := map[string]struct {
-		addr          string
-		isFdConsuming bool
-	}{
-		"tcp": {
-			addr:          "/ip4/127.0.0.1/tcp/20",
-			isFdConsuming: true,
-		},
-		"quic": {
-			addr:          "/ip4/127.0.0.1/udp/0/quic",
-			isFdConsuming: false,
-		},
-		"addr-without-registered-transport": {
-			addr:          "/ip4/127.0.0.1/tcp/20/ws",
-			isFdConsuming: true,
-		},
-		"relay-tcp": {
-			addr:          fmt.Sprintf("/ip4/127.0.0.1/tcp/20/p2p-circuit/p2p/%s", test.RandPeerIDFatal(t)),
-			isFdConsuming: true,
-		},
-		"relay-quic": {
-			addr:          fmt.Sprintf("/ip4/127.0.0.1/udp/20/quic/p2p-circuit/p2p/%s", test.RandPeerIDFatal(t)),
-			isFdConsuming: false,
-		},
-		"relay-without-serveraddr": {
-			addr:          fmt.Sprintf("/p2p-circuit/p2p/%s", test.RandPeerIDFatal(t)),
-			isFdConsuming: true,
-		},
-		"relay-without-registered-transport-server": {
-			addr:          fmt.Sprintf("/ip4/127.0.0.1/tcp/20/ws/p2p-circuit/p2p/%s", test.RandPeerIDFatal(t)),
-			isFdConsuming: true,
-		},
-	}
-
-	ctx := context.Background()
-	sw := GenSwarm(t, ctx)
-	sk := sw.Peerstore().PrivKey(sw.LocalPeer())
-	require.NotNil(t, sk)
-
-	for name := range tcs {
-		maddr, err := ma.NewMultiaddr(tcs[name].addr)
-		require.NoError(t, err, name)
-		require.Equal(t, tcs[name].isFdConsuming, sw.IsFdConsumingAddr(maddr), name)
 	}
 }
